Treat times as UTC
This is not actually accurate but it makes live a lot simpler down the
line.

See https://www.pivotaltracker.com/story/show/62779118 for details.

This commit also changes the `data_id` function to not include the
dimension values as part of the id. This is because they are not passed
through to backdrop and therefore can't be used to rebuild the id at a
later stage if needed. Also, records are never identified by their
dimension values within a single bucked (they would be in separate
buckets).