Don't fail the build if we fail on Python 2.6
Copying what we now do in https://github.com/alphagov/backdrop, we still
run the tests on Python 2.6 but they are allowed to fail without failing
the whole build.

The Python 2.7 features which prompted this are ``assert_is_instance`` from
``nose.tools`` and ``"{}".format(...)`` - the former in particular is a
very useful test method.