Allow id field to be defined
In some situations we want the id of each row to be a known field value,
this is so that we can overwrite existing values with new values.
Dimensions seemed the only reasonable set of fields to pick out of as
deriving an id from the metric you are measuring seemed pointless for
this purpose.

IDs are base64'd such that no characters backdrop doesn't like in IDs
can end up there.